~ Рефакторинг.

+++ TileExplorer/Database/Database.Models.cs
@@ -155,13 +155,13 @@ namespace TileExplorer
 
                 [DisplayName("Скорость")]
                 [Write(false)]
-                public float AverageSpeed { get => gpx.AverageSpeed; }
+                public double AverageSpeed { get => gpx.AverageSpeed; }
 
                 [DisplayName("Подъём")]
-                public float EleAscent { get => gpx.EleAscent; set => gpx.EleAscent = value; }
+                public double EleAscent { get => gpx.EleAscent; set => gpx.EleAscent = value; }
 
                 [DisplayName("Спуск")]
-                public float EleDescent { get => gpx.EleDescent; set => gpx.EleDescent = value; }
+                public double EleDescent { get => gpx.EleDescent; set => gpx.EleDescent = value; }
 
                 [Write(false)]
                 public List<TrackPoint> TrackPoints { get; set; } = null;
